11.3.Twig Block Tricks{Creating a Sub-Layout}
Creating a Sub-Layout
In the templates/ directory, create a new file: content_base.html.twig. What's cool is that, we can extend the normal base.html.twig and then just add the extra markup we need.
To do that, override block body just like we would in a normal template. Then, steal the first four divs: these are the divs that really give us the structure. Paste them here, and type a ton of closing div tags.
Next, and here's the key, in the middle, which is where we want the content to go, create a new block called content_body and {% endblock %}.
I just invented that name.
And, that's it! Let's go use it! In index.html.twig, change the extends to content_base.html.twig.
Now, we do not want to override the block body. Nope, we want to override content_body. Thanks to this, we should get the normal base layout plus the extra markup from content_base.html.twig.
Remove the 4 divs, their closing tags, then clean things up a bit.
Ok, try it! Go back to the admin tab and refresh. Yay! A nice layout with no work. Repeat this in show.html.twig: extend content_base.html.twig, change the block to content_body, and remove the 4 divs on top, the 4 closing tags on the bottom and... un-indent a few times so this looks decent.
And unless we forgot something... nope! It still looks perfect! We now have a super easy way to create new pages.