11.4.Twig Block Tricks{Adding a Custom Class to One Template}
Adding a Custom Class to One Template
Except... there's one small design change I want to make... but I want to make it to the comment admin section only. Our project already has a public/css/styles.css file. On your browser, "Inspect Element" on the white box. One of the CSS classes in styles.css is show-article-container-border-green.
If you add this, you get a nice green border on top! And according to our hard-working and very particular design team, they want this on the "Manage Comments" page, but they do not want it on the article page. Apparently some of our alien readers associate the color green with fictional, untrustworthy content. Well, we can't have that!
But, dang! This ruins everything! The class needs to live on the show-article-container div... but that lives inside content_base.html.twig. How can we change this for only one of the children templates?
The answer is... drumroll... blocks! Blocks are almost always the answer when you need to do cool things with Twig inheritance.
In content_base.html.twig, surround all of the classes with a new block: call it content_class. After the classes, use endblock.
This defines a new block that has default content. If nobody overrides the block, it will have all of these classes. But, in the comments template, we can override this: {% block content_class %}. But, we don't really want to fully replace it: we want to add to the block's content. No problem: use {{ parent() }} to print the existing classes, then show-article-container-border-green with {% endblock %}
I love it! To make sure I'm not lying: find your browser and refresh the manage comments page. Looks great! But the article show page... nope! No green border: our alien readers will continue to trust us.
Next, let's finish our comment admin page by listing them in a table. And, we'll install a cool library called "Twig Extension".