12.1.The Twig Extensions Library
Let's bring this section to life by listing all of the comments in the system. Ah man, with all our tools, this is going to be really easy! First, to query for the comments, add the CommentRepository $repository argument.
Then, $comments = $repository->, and we could use findAll(), but I'll use findBy() passing this an empty array, then 'createdAt' => 'DESC' so that we get the newest comments on top.
Clear out the render() variables: we only need to pass one: comments set to $comments.
Perfect! Next, to the template! Below the h1, I'll paste the beginning of a table that has some Bootstrap classes and headers for the article name, author, the comment itself and when it was created.
No problem! In the tbody, let's loop: for comment in comments, and {% endfor %}.
Add the <tr>, then let's print some data! In the first td, we need the article name. But, to make it more awesome, let's make this a link to the article. Add the a tag with href="", but keep that blank for a moment. Inside, hmm, we have a Comment object, but we want to print the article's title. No problem! We can use our relationship: comment.article - that gets us to the Article object - then .title.
For the href, use the path() function from Twig. Here, we need the name of the route that we want to link to. Open ArticleController. Ah! There it is: name="article_show
Close that and, back in the template, use article_show. This route needs a slug parameter so add that, set to comment.article.slug.
Dang, those relationships are handy!
Let's keep going! Add another td and print comment.authorName.
Give the next td a style="width: 20%" so it doesn't get too big. Then, print comment.content.
Finally, add a td with comment.createdAt|ago.
Cool! Let's see if we made any mistakes. Find your browser, refresh and... boom! A big, beautiful list of all of the comments on the site. Oh, but eventually on production, this will be a huge number of results. Let's put it on our todo list to add pagination.