13.1.Request Object & Query OR Logic
Because astronauts love to debate news, our site will have a lot of comments on production. So, let's add a search box above this table so we can find things quickly.
Open the template and, on top, I'm going to paste in a simple HTML form
We're not going to use Symfony's form system because, first, we haven't learned about it yet, and second, this is a super simple form: Symfony's form system wouldn't help us much anyways.
Ok! Check this out: the form has one input field whose name is q, and a button at the bottom. Notice that the form has no action=: this means that the form will submit right back to this same URL. It also has no method=, which means it will submit with a GET request instead of POST, which is exactly what you want for a search or filter form.
Let's see what it looks like: find your browser and refresh. Nice! Search for "ipsam" and hit enter. No, the search won't magically work yet. But, we can see the ?q= at the end of the URL.