13.3.Request Object & Query OR Logic{A Custom Query with OR Logic}
A Custom Query with OR Logic
Now that we have the search term, we need to use that to make a custom query. So, sadly, we cannot use findBy() anymore: it's not smart enough to do queries that use the LIKE keyword. No worries: inside CommentRepository, add a public function called findAllWithSearch(). Give this a nullable string argument called $term
I'm making this nullable because, for convenience, I want to allow this method to be called with a null term, and we'll be smart enough to just return everything.
Above the method, add some PHP doc: this will @return an array of Comment objects.
Ok: we already know how to write custom queries: $this->createQueryBuilder() with an alias of c.
Then, if a $term is passed, we need a WHERE clause. But, here's the tricky part: I want to search for the term on a couple of fields: I want WHERE content LIKE $term OR authorName LIKE $term.
How can we do this? Hmm, the QueryBuilder apparently has an orWhere() method. Perfect, right? No! Surprise, I never use this method. Why? Imagine a complex query with various levels of AND clauses mixed with OR clauses and parenthesis. With a complex query like this, you would need to be very careful to use the parenthesis in just the right places. One mistake could lead to an OR causing many more results to be returned than you expect!
To best handle this in Doctrine, always use andWhere() and put all the OR logic right inside: c.content LIKE :term OR c.authorName LIKE :term. On the next line, set term to, this looks a little odd, '%'.$term.'%'.
By putting this all inside andWhere() - instead of orWhere() - all of that logic will be surrounded by a parenthesis. Later, if we add another andWhere(), it'll logically group together properly.
Finally, in all cases, we want to return $qb->orderBy('c.createdAt', 'DESC') and ->getQuery()->getResult().
Remember, getResult() returns an array of results, and getOneOrNullResult() returns just one row.
Phew! That looks great! Go back to the controller. Use that method: $comments = $repository->findAllWithSearch() passing it $q.
Moment of truth! First, remove the ?q= from the URL. Ok, everything looks good. Now search for something very specific, like, ahem, reprehenderit. And, yes! A much smaller result. Try an author: Ernie: got it!

Woo! This is great! But, we can do more! Next, let's learn about a Twig global variable that can help us fill in this input box when we search. Then, it's finally time to add a join to our custom query.