14.2.Query Joins & Solving the N+1 Problem{The Twig For-Else Feature}
The Twig For-Else Feature
Oh, by the way, here's a fun Twig feature. When we get zero results, we should probably print a nice message. On a Twig for loop, you can put an else at the end. Add a <td colspan=4>, a centering class, and: No comments found.
Go back and try it! It works! Pff, except for my not-awesome styling skills. Use text-center.
That's better.