14.4.Query Joins & Solving the N+1 Problem{Solving the N+1 (Extra Queries) Problem}
Solving the N+1 (Extra Queries) Problem

But at this point, the solution to the N+1 problem is dead simple. Go back to CommentRepository and put ->addSelect('a')
When you create a QueryBuilder from inside a repository, that QueryBuilder automatically knows to select from its own table, so, from c. With this line, we're telling the QueryBuilder to select all of the comment columns and all of the article columns.
Try it: head back and refresh. It still works! But, yes! We're down to just one query. Go check it out: yep! It selects everything from comment and article.
The moral of the story is this: if your page has a lot of queries because Doctrine is making extra queries across a relationship, just join over that relationship and use addSelect() to fetch all the data you need at once.
But... there is one confusing thing about this. We're now selecting all of the comment data and all of the article data. But... you'll notice, the page still works! What I mean is, even though we're suddenly selecting more data, our findAllWithSearch() method still returns exactly what it did before: it returns a array of Comment objects. It does not, for example, now return Comment and Article objects.
Instead, Doctrine takes that extra article data and stores it in the background for later. But, the new addSelect() does not affect the return value. That's way different than using raw SQL.
It's now time to cross off a todo from earlier: let's add pagination