16.2.The 4 (2?) Possible Relation Types{Building the Tag Entity: php bin/console make:entity}
Building the Tag Entity
Let's create the new Tag entity class first. Find your terminal and run:

php bin/console make:entity
Name it Tag and give it two properties: name, as a string and slug also as a string, so that we can use the tag in a URL later.