16.3.The 4 (2?) Possible Relation Types{Building the Tag Entity}
Cool! Before generating the migration, open the new class. No surprises: name and slug. At the top, use our favorite TimestampableEntity trait.
And, just like we did in Article, configure the slug to generate automatically. Copy the slug annotation and paste that above the slug property.
Oh, but we need a use statement for the annotation. An easy way to add it is to temporarily type @Slug on the next line and hit tab to auto-complete it. Then, delete it: that was enough to make sure the use statement was added on top.
Let's also make the slug column unique.