16.6.The 4 (2?) Possible Relation Types{Building the Tag Entity}
First, extend BaseFixture, rename load() to loadData() and make it protected.
We also don't need this use statement anymore. Call our trusty $this->createMany() to create 10 tags.
For the name, use $tag->setName() with $this->faker->realText() and 20, to get about that many characters.
We could use $this->faker->word to get a random word, but that word would be in Latin. The realText() method will give us a few words, actually, but they will sound, at least "kinda" real.

And, that's all we need! To make sure it works, run:

php bin/console doctrine:fixtures:load
We are ready! Article entity, check! Tag entity, check, check! It's time to create the ManyToMany relationship.