17.2.ManyToMany Relationship{Generating the Migration: php bin/console make:migration}
Generating the Migration
Enough of that! Let's generate the migration:

php bin/console make:migration
Cool! Go open that file
Woh! It creates a new table! Of course! That's how you model a ManyToMany relationship in a relational database. It creates an article_tag table with only two fields: article_id and tag_id.

This is very different than anything we've seen so far with Doctrine. This is the first time - and really, the only time - that you will have a table in the database, that has no direct entity class. This table is created magically by Doctrine to help us relate tags and articles. And, as we'll see next, Doctrine will also automatically insert and delete records from this table as we add and remove tags from an article.

Now, run the migration:

php bin/console doctrine:migrations:migrate
Let's go tag some articles!