18.2.Saving a ManyToMany Relation + Joins
Next, in ArticleFixtures, this is where we'll set the relationship. And that means, we need to make sure that TagFixture is loaded first so that the tags actually exist. At the top, add implements DependentFixtureInterface.
Then, I'll go to the "Code"->"Generate" menu - or Command+N on a Mac - select "Implement Methods" and choose getDependencies(). We now depend on TagFixture::class.
Above, let's first get some tag objects: $tags = $this->getRandomReferences() and pass it Tag::class, and then, let's fetch $this->faker->numberBetween() zero and five. So, find 0 to 5 random tags.
And just to make sure you that this does give us Tag objects, dump($tags) and die. Now, find your terminal and run:

php bin/console doctrine:fixtures:load
Perfect! These are Tag objects. Oh, by the way, sometimes you may notice that your entity's class name is prefixed by this weird Proxies stuff. When you see that, ignore it. A "Proxy" is a special class that Doctrine generates and sometimes wraps around your real entity objects. Doctrine does this so that it can perform its relationship lazy-loading magic.
Actually, check this out: it looks like all the data on this Tag is null! But, that's a lie! As soon as you reference any data on that Tag, Doctrine will query for the data and fill it in. That's lazy-loading in action.