18.3.Saving a ManyToMany Relation + Joins{Explaining Proxies}
Explaining Proxies

Let me show you: add a foreach over $tags as $tag.
To help PhpStorm, I'll use some inline PHPDoc to tell it that  is an array of Tag objects.

Inside the loop, just say $tag->getName()
I know, that looks weird: we're calling a method but not using it! But, calling this method is enough to make Doctrine query for the tag's real data. Below, dump($tag) and die after the loop.

Load the fixtures again:

php bin/console doctrine:fixtures:load
Boom! We have data!

Anyways, this is the proxy system in action. You need to know what it is because you will see it from time-to-time. But mostly, it should be completely invisible: don't think about it.